+++ test/yj_vertex_set_path_composite.test.cpp
@@ -19,8 +19,8 @@ int main() {
 
   std::vector<composite_monoid<mi>> f(n);
   for (auto& fi: f) {
-    intmax_t a, b;
-    scanf("%jd %jd", &a, &b);
+    int a, b;
+    scanf("%d %d", &a, &b);
     fi = {a, b};
   }
 
@@ -39,14 +39,14 @@ int main() {
 
     if (t == 0) {
       size_t p;
-      intmax_t c, d;
-      scanf("%zu %jd %jd", &p, &c, &d);
+      int c, d;
+      scanf("%zu %d %d", &p, &c, &d);
       f[p] = {c, d};
       g.set(p, f[p]);
     } else if (t == 1) {
       size_t u, v;
-      intmax_t x;
-      scanf("%zu %zu %jd", &u, &v, &x);
+      int x;
+      scanf("%zu %zu %d", &u, &v, &x);
       printf("%d\n", g.fold(u, v)(x).get());
     }
   }
